Fix unix-specific child killing
